fix/replace hooks
